+++ datasets/d-keep-code_davinci_001_temp_0.2/HumanEval_9_rolling_max.d
@@ -1,5 +1,5 @@
-import std.typecons;
 import std.math;
+import std.typecons;
 /*
  From a given list of integers, generate a list of rolling maximum element found until given moment
     in the sequence.
@@ -7,12 +7,11 @@ import std.math;
     [1, 2, 3, 3, 3, 4, 4]
     
 */
-int[] rolling_max(int[] numbers) 
+long[] rolling_max(long[] numbers) 
 {
-    int[] result = new int[numbers.length];
+    long[] result = new long[numbers.length];
     result[0] = numbers[0];
-    for (int i = 1; i < numbers.length; i++)
-    {
+    for (int i = 1; i < numbers.length; i++) {
         result[i] = std.math.max(numbers[i], result[i - 1]);
     }
     return result;
